// src/utils/Enums/Elements.enum.ts
export enum ZFill {
  clear = 'clear',
  default = 'default',
  outline = 'outline',
  solid = 'solid'
}

export enum ZLabelPosition {
  fixed = 'fixed',
  floating = 'floating',
  stacked = 'stacked'
}

export enum ZColorEnum {
  primary = 'primary',
  secondary = 'secondary',
  danger = 'danger',
  success = 'success',
  light = 'light',
  white = 'white',
  dark = 'dark',
  medium = 'medium',
  tertiary = 'tertiary',
  warning = 'warning',
  shadowedPlum = 'shadowedPlum'
}